Move guest_user to `guest` component.
