+++ app/api/v1/api.py
@@ -1,12 +1,11 @@
 from fastapi import APIRouter
 
 from app.api.v1.endpoints import geocoding
-from app.api.v1.endpoints import guest_user
 from app.api.v1.endpoints import sessions
-from app.components import locations, oauth, zones, auth, organizations, users
+from app.components import locations, oauth, zones, auth, organizations, users, guests
 
 api_router = APIRouter()
-
+# TODO: Consider moving prefixes and tags to router declarations.
 api_router.include_router(auth.routes.router, prefix="/auth", tags=["auth"])
 api_router.include_router(
     locations.routes.router, prefix="/locations", tags=["locations"]
@@ -18,5 +17,5 @@ api_router.include_router(
 api_router.include_router(sessions.router, prefix="/sessions", tags=["sessions"])
 api_router.include_router(oauth.routes.router, prefix="/oauth", tags=["oauth"])
 api_router.include_router(zones.routes.router, prefix="/zones", tags=["zones"])
-api_router.include_router(guest_user.router, prefix="/guest", tags=["guest-user"])
+api_router.include_router(guests.routes.router, prefix="/guest", tags=["guest-user"])
 api_router.include_router(geocoding.router, prefix="/geocode", tags=["geocoding"])
