+++ app/components/guests/models.py
@@ -1,5 +1,4 @@
 from sqlalchemy import Column, Integer, String, DateTime
-
 from sqlalchemy.orm import relationship
 
 from app.db.base_class import Base
