+++ app/components/locations/models.py
@@ -3,9 +3,9 @@ from sqlalchemy import Column, Integer, String, DateTime, Float, ForeignKey
 from sqlalchemy.dialects.postgresql import JSONB
 from sqlalchemy.orm import relationship
 
+from app.components.guests.models import GuestUser
 from app.db.base_class import Base
 from app.db.utc_convertation import utcnow
-from app.models.guest_user import GuestUser
 
 status_list = {1: "Awaiting review", 2: "Awaiting approval", 3: "Approved"}
