+++ app/db/base.py
@@ -1,4 +1,5 @@
 from app.components.geospatial.models import GeospatialIndex
+from app.components.guests.models import GuestUser
 from app.components.locations.models import Location
 from app.components.oauth.models import OauthScope, OauthRole, association_table
 from app.components.organizations.models import Organization
@@ -7,7 +8,6 @@ from app.components.users.models import User
 from app.components.zones.models import Zone
 from app.db.base_class import Base
 from app.models.changelog import ChangeLog
-from app.models.guest_user import GuestUser
 from app.models.sessionhistory import SessionHistory
 
 __all__ = [
