+++ app/schemas/__init__.py
@@ -1,7 +1,6 @@
 from .changelog import ChangelogOut
 from .enums import *
 from .geocoding import *
-from .guest_user import LocationRequestOtp
 from .roles import UserRole
 from .session import UserSession
 from .session import UserSession
